// datahub/src/types/Transaction.ts
export interface BankingRecord {
  customer_id: string;
  first_name: string;
  last_name: string;
  age: string;
  gender: string;
  address: string;
  city: string;
  contact_number: string;
  email: string;
  account_type: string;
  account_balance: string;
  date_of_account_opening: string;
  last_transaction_date: string;
  transaction_id: string;
  transaction_date: string;
  transaction_type: string;
  transaction_amount: string;
  account_balance_after_transaction: string;
  branch_id: string;
  loan_id: string;
  loan_amount: string;
  loa_type: string;
  interest_rate: string;
  loan_term: string;
  approval_rejection_date: string;
  loan_status: string;
  card_id: string;
  card_type: string;
  credit_limit: string;
  credit_card_balance: string;
  minimum_payment_due: string;
  payment_due_date: string;
  last_credit_card_payment_date: string;
  rewards_points: string;
  feedback_id: string;
  feedback_date: string;
  feedback_type: string;
  resolution_status: string;
  resolution_date: string;
  anomaly: string;
}